move text utils to dedicated utils file
